+++ sycl/source/detail/xpti_registry.cpp
@@ -26,8 +26,7 @@ xpti::trace_event_data_t *XPTIRegistry::createTraceEvent(
                                          const_cast<void *>(FuncPtr));
   xpti::payload_t Payload(Name.c_str(),
                           (CodeLoc.fileName() ? CodeLoc.fileName() : ""),
-                          static_cast<int>(CodeLoc.lineNumber()),
-                          static_cast<int>(CodeLoc.columnNumber()), Obj);
+                          CodeLoc.lineNumber(), CodeLoc.columnNumber(), Obj);
 
   // Calls could be at different user-code locations; We create a new event
   // based on the code location info and if this has been seen before, a
